Initial proposal to support Object Mapping (or Binding) (#73)
* Initial proposal to support Object Mapping (or Binding)

* Structure and format specification

This is a first pass at creating a specification for structure and format. Created in an AsciIDoc file for convenience, but can easily be split and moved to API JavaDoc on a class or package, or made part of `overview.html` as desired.

Tried to fill in things that we were in agreement on. Still have to mine earlier meeting minutes and discussion threads to piece together the rest.

* Removed ConfigMapping Appendix

Co-authored-by: David M. Lloyd <david.lloyd@redhat.com>